# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "LumenMedia.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:15080/"
    }
  ],
  "paths": {
...
    "/api/v1/items/{id}": {
      "get": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDetail"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/series/{id}/seasons": {
      "get": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfSeasonDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfSeasonDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfSeasonDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/seasons/{id}/episodes": {
      "get": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfEpisodeSummary"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfEpisodeSummary"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfEpisodeSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/episodes/{id}": {
      "get": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeDetail"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search": {
      "get": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/items/{id}/artwork/{kind}": {
      "get": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ArtworkKind"
            }
          },
          {
            "name": "w",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "h",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "quality",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/items/{id}/match-candidates": {
      "get": {
        "tags": [
          "Metadata"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetadataMatchCandidateDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetadataMatchCandidateDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetadataMatchCandidateDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/items/{id}/match": {
      "post": {
        "tags": [
          "Metadata"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/items/{id}/refresh-metadata": {
      "post": {
        "tags": [
          "Metadata"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/items/{id}/metadata": {
      "patch": {
        "tags": [
          "Metadata"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemMetadataRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemMetadataRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemMetadataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArtworkKind": {
        "enum": [
          "Poster",
          "Backdrop",
          "Logo",
          "Thumb",
          "Banner"
        ],
        "type": "string"
      },
      "ArtworkUrls": {
        "type": "object",
        "properties": {
          "poster": {
            "type": [
              "null",
              "string"
            ]
          },
          "backdrop": {
            "type": [
              "null",
              "string"
            ]
          },
          "logo": {
            "type": [
              "null",
              "string"
            ]
          },
          "thumb": {
            "type": [
              "null",
              "string"
            ]
          },
          "banner": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "EpisodeDetail": {
        "required": [
          "id",
          "seriesId",
          "seasonId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/MediaKind"
          },
          "seriesId": {
            "type": "string",
            "format": "uuid"
          },
          "seasonId": {
            "type": "string",
            "format": "uuid"
          },
          "seasonNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "episodeNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "overview": {
            "type": [
              "null",
              "string"
            ]
          },
          "airDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "runtimeMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "artwork": {
            "$ref": "#/components/schemas/ArtworkUrls"
          },
          "mediaSources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaSourceDto"
            }
          },
          "userData": {
            "$ref": "#/components/schemas/UserDataDto"
          }
        }
      },
...
      "ExternalIds": {
        "type": "object",
        "properties": {
          "tmdb": {
            "type": [
              "null",
              "string"
            ]
          },
          "tvdb": {
            "type": [
              "null",
              "string"
            ]
          },
          "imdb": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "JobDto": {
        "required": [
          "id",
          "type",
          "state"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "$ref": "#/components/schemas/JobType"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "progress": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          },
          "libraryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "startedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "finishedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "error": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "JobType": {
        "enum": [
          "ScanLibrary",
          "ImportFile",
          "FetchMetadata",
          "GenerateArtwork",
          "CleanupTranscodes"
        ],
        "type": "string"
      },
...
      "MatchRequest": {
        "required": [
          "provider",
          "providerId"
        ],
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "providerId": {
            "type": "string"
          }
        }
      },
...
      "MediaKind": {
        "enum": [
          "Movie",
          "Series",
          "Episode"
        ],
        "type": "string"
      },
...
      "MetadataMatchCandidateDto": {
        "required": [
          "provider",
          "providerId",
          "title"
        ],
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "providerId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "year": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "score": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "PagedResultOfEpisodeSummary": {
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EpisodeSummary"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PagedResultOfSeasonDto": {
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeasonDto"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PersonDto": {
        "required": [
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "role": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": "string"
          },
          "order": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "thumb": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SearchResponse": {
        "type": "object",
        "properties": {
          "movies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaItemSummary"
            }
          },
          "series": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaItemSummary"
            }
          },
          "episodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EpisodeSummary"
            }
          }
        }
      },
      "SeasonDto": {
        "required": [
          "id",
          "seriesId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "seriesId": {
            "type": "string",
            "format": "uuid"
          },
          "seasonNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "episodeCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "artwork": {
            "$ref": "#/components/schemas/ArtworkUrls"
          }
        }
      },
...
      "SeriesDetail": {
        "required": [
          "id",
          "title",
          "libraryId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/MediaKind"
          },
          "title": {
            "type": "string"
          },
          "originalTitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "year": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "endYear": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "status": {
            "type": [
              "null",
              "string"
            ]
          },
          "overview": {
            "type": [
              "null",
              "string"
            ]
          },
          "communityRating": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "officialRating": {
            "type": [
              "null",
              "string"
            ]
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "people": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersonDto"
            }
          },
          "trailerUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "externalIds": {
            "$ref": "#/components/schemas/ExternalIds"
          },
          "metadataLocked": {
            "type": "boolean"
          },
          "seasonCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "episodeCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "artwork": {
            "$ref": "#/components/schemas/ArtworkUrls"
          },
          "userData": {
            "$ref": "#/components/schemas/SeriesUserData"
          },
          "libraryId": {
            "type": "string",
            "format": "uuid"
          },
          "addedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SeriesUserData": {
        "type": "object",
        "properties": {
          "unwatchedEpisodeCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nextUp": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EpisodeSummary"
              }
            ]
          }
        }
      },
...
      "UpdateItemMetadataRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "originalTitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "year": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "overview": {
            "type": [
              "null",
              "string"
            ]
          },
          "tagline": {
            "type": [
              "null",
              "string"
            ]
          },
          "communityRating": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "officialRating": {
            "type": [
              "null",
              "string"
            ]
          },
          "metadataLocked": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
...
    }
  },
...
}